::Constants: fix usage of strict/warnings
"use strict" and "use warnings" are lexical. "package" doesn't define a
new lexical context. So using "use strict; use warnings;" just once at
the top of the file is enough.

In a similar way the true value (usually "1;") required at the end of a
module only applies to the last statement of the *file*, not as the last
statement of a package (because such a thing doesn't exist anyway).