Fix handling for format declaration
If `format` is used at non declaration context (for example: key of a hash),
it makes incapable of tokenizing action. It breaks structure of tokens.

I'll show an example code to reproduce;

```perl
my $foo = {
    format => 1,  # <= here's format is misrecognized as a FormatDecl
};

my $bar =  # <= code which is after assign operator will be broken
  "asdf";
1;
```

When tokenizing this code, result of it will be broken because `format`
which is a key of hash is misrecognized as a `FormatDecl`.
So I fixed it.